refactor(timeline): split day-jump into its own hook
useTimelineScrollSync now owns only the passive URL <-> scroll sync;
imperative jumpTo moves to a new useTimelineJump hook.
